work on image loading system and basic ui work
images of the maps now load, but took alot longer than expected due to dice's inconsistent naming schemes for the loading images.

the loading screen now hides when the player is fully connected (i.e. spawn screen shows), however the loading screen will not show again in a session. will be fixed in the future

preliminary UI work. it shows the map name and the current game mode. I want to show the server information and the more details (hz and such) but thats later